requirements.txt
Flask==3.1.1
flask-cors==6.0.0
google-genai
python-dotenv